src/main/java/org/example/service/UserService.java
package org.example.service;

import lombok.Data;
import org.example.dao.UserDao;

import java.util.*;
import java.util.function.Function;

import static java.util.stream.Collectors.toMap;

/**
 * Hello world!
 */
public class UserService {

    private final UserDao userDao;

    public UserService(UserDao userDao) {
        this.userDao = userDao;
    }

    public boolean delete(Integer userId){
        return userDao.delete(userId);
    }

    private final List<User> users = new ArrayList<>();

    public List<User> getAll() {
        return users;
    }

    public void add(User... users) {
        Collections.addAll(this.users, users);
    }

    public Optional<User> login(String username, String password) {
        if(username == null || password == null)
            throw new IllegalArgumentException("username or password is null");
        return users.stream()
                .filter(user -> user.getUsername().equals(username))
                .filter(user -> user.getPassword().equals(password))
                .findFirst();
    }

    public Map<Integer, User> getAllConvertedById() {
        return users.stream()
                .collect(toMap(User::getId, Function.identity()));
    }
}
